+++ test/test_wollok.py
@@ -27,8 +27,12 @@ class WollokKernelTests(jkt.KernelTests):
     # code samples
     code_execute_result = [
         {
-            "code": "6",
-            "result": "e7f6c011776e8db7cd330b54174fd76f7d0216b612387a5ffcfb81e6f0919683"
+            "code": "6.even()",
+            "result": "✓ true"
+        },
+        {
+            "code": "3.mayor()",
+            "result": "✗ Evaluation Error!\n  3 does not understand mayor()"
         }
     ]